server/requirements.txt
django
django-admin-sortable2
django-ckeditor-updated
django-positions
djangorestframework
drf-extensions
django-filter
django-taggit
sorl-thumbnail>=11.12.1b
pillow
model_mommy
django-cors-headers
django-custom-user
